requirements.txt
pdf2image~=1.17.0
pytesseract~=0.3.13
pandas~=2.2.3
streamlit~=1.45.1
pillow~=11.2.1
streamlit
opencv-python
regex